Move source actions to source attributes view mode
