+++ src/components/nodes/DataSourceDrawer.svelte
@@ -9,11 +9,11 @@
     Input,
     Textarea,
   } from "flowbite-svelte";
+  import { map } from "../Map.svelte";
 
   let handler, rows;
   export let sourceNode,
     hiddenBool = false;
-  let position;
 
   // Attribute table
   // Create a new data handler for the table
@@ -27,17 +27,20 @@
     string = string.replace("_", " ");
     return string;
   }
-  // Update position of node in store
-  function updatePosition() {
-    if (position != null) {
-      uploadedSources.update((nodes) => {
-        let index = nodes.findIndex(
-          (node) => node.fileName == sourceNode.fileName
-        );
-        nodes[index].position = position;
-        return nodes;
-      });
+
+  async function removeSource(layer) {
+    if (layer["type"] == "Spatial") {
+      map.removeLayer(layer["name"]);
+      map.removeSource(layer["fileName"]);
+    } else if (layer["type"] == "PDF" || layer["type"] == "Link") {
+      // // destroy the iframe
+      // layer.container.innerHTML = "";
+      // // set display css property to none
+      // layer.container.style.display = "none";
     }
+    uploadedSources.update((sources) => {
+      return sources.filter((source) => source["name"] != layer["name"]);
+    });
   }
 </script>
 
@@ -55,56 +58,63 @@
         class="mb-4 absolute right-7 dark:text-white"
       />
       <h3 class="text-3xl my-3">{insertSpaces(sourceNode.fileName)}</h3>
-      <ul class="text-lg">
-        <li>Type: {sourceNode.type}</li>
+      <section class="text-2md mb-4">
+        <p>Type: {sourceNode.type}</p>
         {#if sourceNode.type === "Spatial"}
-          <li>Geometry: {sourceNode.geometry}</li>
+          <p>Geometry: {sourceNode.geometry}</p>
         {/if}
+      </section>
+      <section id="actions" class="mb-3">
+        <Button
+          size="xs"
+          class="px-1 py-0 m-0 rounded-sm"
+          on:click={() => {
+            removeSource(sourceNode);
+          }}
+        >
+          Remove Source</Button
+        >
+      </section>
+      {#if sourceNode.attributes}
+        <section>
+          <Datatable {handler} class="text-gray-700">
+            <table>
+              <thead>
+                <Th {handler} orderBy="name">Name</Th>
+                <Th {handler} orderBy="type">DataType</Th>
+                <Th {handler} orderBy="range">Range</Th>
+              </thead>
+              <tbody>
+                {#each $rows as row}
+                  <tr>
+                    <td>{@html row.name}</td>
+                    <td>{@html row.dataType}</td>
+                    <td>{@html row.range}</td>
+                  </tr>
+                {/each}
+              </tbody>
+            </table>
+          </Datatable>
+        </section>
+      {/if}
 
-        {#if sourceNode.attributes}
-          <section>
-            <Datatable {handler} class="text-gray-700">
-              <table>
-                <thead>
-                  <Th {handler} orderBy="name">Name</Th>
-                  <Th {handler} orderBy="type">DataType</Th>
-                  <Th {handler} orderBy="range">Range</Th>
-                </thead>
-                <tbody>
-                  {#each $rows as row}
-                    <tr>
-                      <td>{@html row.name}</td>
-                      <td>{@html row.dataType}</td>
-                      <td>{@html row.dataType}</td>
-                      <td>{@html row.range}</td>
-                    </tr>
-                  {/each}
-                </tbody>
-              </table>
-            </Datatable>
-          </section>
-        {/if}
-        <li>
-          {#if sourceNode.source}
-            Source: {sourceNode.source}
-          {:else}
-            <Label for="Source" class="!text-black text-lg ">Source</Label>
-            <Input
-              class="bg-inputField-200"
-              placeholder="Provide a source link"
-              name="citation"
-            />
-          {/if}
-        </li>
-        <li>
-          <Label for="citation" class="!text-black text-lg">Citation</Label>
-          <Textarea
-            class="bg-inputField-200"
-            placeholder="Please provide a citation before sharing"
-            name="citation"
-          />
-        </li>
-      </ul>
+      {#if sourceNode.source}
+        Source: {sourceNode.source}
+      {:else}
+        <Label for="Source" class="!text-black text-lg ">Source</Label>
+        <Input
+          class="bg-inputField-200"
+          placeholder="Provide a source link"
+          name="citation"
+        />
+      {/if}
+
+      <Label for="citation" class="!text-black text-lg">Citation</Label>
+      <Textarea
+        class="bg-inputField-200"
+        placeholder="Please provide a citation before sharing"
+        name="citation"
+      />
     </div>
   </div>
 </Drawer>
